[JW] tried to merge 1st page to 3rd page
