--- flutter/safeway_direction/lib/iPhoneXXSpage221/iPhoneXXSintro1.dart
@@ -1,162 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_svg/flutter_svg.dart';
-
-class iPhoneXXSintro1 extends StatelessWidget {
-  iPhoneXXSintro1({
-    Key key,
-  }) : super(key: key);
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Stack(
-        children: <Widget>[
-          Transform.translate(
-            offset: Offset(0.0, -6.0),
-            child: Stack(
-              children: <Widget>[
-                Stack(
-                  children: <Widget>[
-                    Transform.translate(
-                      offset: Offset(73.0, 493.83),
-                      child: SizedBox(
-                        width: 230.0,
-                        child: Text(
-                          'KIDSKEEPER',
-                          style: TextStyle(
-                            fontFamily: 'BM JUA_TTF',
-                            fontSize: 35,
-                            color: const Color(0xffee8700),
-                            height: 0.6571428571428571,
-                          ),
-                          textAlign: TextAlign.center,
-                        ),
-                      ),
-                    ),
-                    Transform.translate(
-                      offset: Offset(13.02, 0.0),
-                      child: Stack(
-                        children: <Widget>[
-                          Transform.translate(
-                            offset: Offset(139.98, 435.0),
-                            child: Container(
-                              width: 70.0,
-                              height: 30.0,
-                              decoration: BoxDecoration(
-                                borderRadius: BorderRadius.all(
-                                    Radius.elliptical(35.0, 15.0)),
-                                color: const Color(0xff684600),
-                              ),
-                            ),
-                          ),
-                          Transform.translate(
-                            offset: Offset(110.05, 276.51),
-                            child: Stack(
-                              children: <Widget>[
-                                Transform.translate(
-                                  offset: Offset(0.0, 0.0),
-                                  child: Stack(
-                                    children: <Widget>[
-                                      Stack(
-                                        children: <Widget>[
-                                          SvgPicture.string(
-                                            _svg_eizxel,
-                                            allowDrawingOutsideViewBox: true,
-                                          ),
-                                        ],
-                                      ),
-                                    ],
-                                  ),
-                                ),
-                                Transform.translate(
-                                  offset: Offset(15.04, 17.03),
-                                  child: SvgPicture.string(
-                                    _svg_htcz4t,
-                                    allowDrawingOutsideViewBox: true,
-                                  ),
-                                ),
-                                Transform.translate(
-                                  offset: Offset(41.53, 44.56),
-                                  child: Stack(
-                                    children: <Widget>[
-                                      Transform.translate(
-                                        offset: Offset(0.0, 0.0),
-                                        child: Stack(
-                                          children: <Widget>[
-                                            Transform.translate(
-                                              offset: Offset(0.0, 0.0),
-                                              child: Stack(
-                                                children: <Widget>[
-                                                  Transform.translate(
-                                                    offset: Offset(0.4, -0.07),
-                                                    child: Container(
-                                                      width: 16.0,
-                                                      height: 27.0,
-                                                      decoration: BoxDecoration(
-                                                        borderRadius:
-                                                            BorderRadius.all(
-                                                                Radius
-                                                                    .elliptical(
-                                                                        8.0,
-                                                                        13.5)),
-                                                        color: const Color(
-                                                            0xffff9100),
-                                                      ),
-                                                    ),
-                                                  ),
-                                                  Transform.translate(
-                                                    offset: Offset(30.4, -0.07),
-                                                    child: Container(
-                                                      width: 16.0,
-                                                      height: 27.0,
-                                                      decoration: BoxDecoration(
-                                                        borderRadius:
-                                                            BorderRadius.all(
-                                                                Radius
-                                                                    .elliptical(
-                                                                        8.0,
-                                                                        13.5)),
-                                                        color: const Color(
-                                                            0xffff9100),
-                                                      ),
-                                                    ),
-                                                  ),
-                                                ],
-                                              ),
-                                            ),
-                                            Transform.translate(
-                                              offset: Offset(4.3, 35.56),
-                                              child: SvgPicture.string(
-                                                _svg_koxa0,
-                                                allowDrawingOutsideViewBox:
-                                                    true,
-                                              ),
-                                            ),
-                                          ],
-                                        ),
-                                      ),
-                                    ],
-                                  ),
-                                ),
-                              ],
-                            ),
-                          ),
-                        ],
-                      ),
-                    ),
-                  ],
-                ),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-const String _svg_eizxel =
-    '<svg viewBox="0.0 0.0 129.9 173.1" ><path transform="translate(-348.67, -280.0)" d="M 413.597412109375 280 C 377.792724609375 280 348.6669921875 309.1290893554688 348.6669921875 344.9304809570312 C 348.6669921875 392.8322448730469 408.5741577148438 449.76611328125 411.1246337890625 452.1685791015625 C 411.8219604492188 452.8202819824219 412.709716796875 453.14794921875 413.597412109375 453.14794921875 C 414.4851684570312 453.14794921875 415.372802734375 452.8202819824219 416.0704956054688 452.1685791015625 C 418.6207275390625 449.76611328125 478.5278930664062 392.8322448730469 478.5278930664062 344.9304809570312 C 478.5278930664062 309.1290893554688 449.402099609375 280 413.597412109375 280 Z M 413.597412109375 381.0030517578125 C 393.7081298828125 381.0030517578125 377.5248413085938 364.81982421875 377.5248413085938 344.9304809570312 C 377.5248413085938 325.0411376953125 393.7084350585938 308.85791015625 413.597412109375 308.85791015625 C 433.486328125 308.85791015625 449.6699829101562 325.0414428710938 449.6699829101562 344.9304809570312 C 449.6699829101562 364.8194885253906 433.48681640625 381.0030517578125 413.597412109375 381.0030517578125 Z" fill="#ff9100" stroke="none" stroke-width="1" stroke-miterlimit="10" stroke-linecap="butt" /></svg>';
-const String _svg_koxa0 =
-    '<svg viewBox="4.3 35.6 38.2 17.9" ><path transform="translate(4.3, 24.06)" d="M 38.23548889160156 11.5 C 38.23548889160156 21.38271522521973 29.67618179321289 29.39424133300781 19.11774444580078 29.39424133300781 C 8.559305191040039 29.39424133300781 0 21.38271522521973 0 11.5 C 16.22097015380859 11.5 34.35827255249023 11.5 38.23548889160156 11.5 Z" fill="#ffc300" stroke="none" stroke-width="1" stroke-miterlimit="4" stroke-linecap="butt" /><path transform="translate(10.56, 44.14)" d="M 0 4.65716552734375 C 3.396240234375 1.7635498046875 7.909912109375 0 12.8642578125 0 C 17.81884765625 0 22.332763671875 1.763671875 25.7291259765625 4.6572265625 C 22.3328857421875 7.55078125 17.8192138671875 9.31439208984375 12.8648681640625 9.31439208984375 C 7.9102783203125 9.31439208984375 3.3963623046875 7.5506591796875 0 4.65716552734375 Z" fill="#fff700" stroke="none" stroke-width="1" stroke-miterlimit="4" stroke-linecap="butt" /></svg>';
-const String _svg_htcz4t =
-    '<svg viewBox="15.0 17.0 100.0 100.0" ><path transform="translate(15.04, 17.03)" d="M 49.99494934082031 0 C 77.60639953613281 0 99.98989868164062 22.38349914550781 99.98989868164062 49.99494934082031 C 99.98989868164062 77.60640716552734 77.60639953613281 99.98989868164062 49.99494934082031 99.98989868164062 C 22.38349914550781 99.98989868164062 0 77.60640716552734 0 49.99494934082031 C 0 22.38349914550781 22.38349914550781 0 49.99494934082031 0 Z" fill="#fffde0" stroke="none" stroke-width="1" stroke-miterlimit="10" stroke-linecap="butt" /></svg>';
