--- flutter/safeway_direction/lib/iPhoneXXSpage221/iPhoneXXSpage221.dart
@@ -1,712 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_svg/flutter_svg.dart';
-import 'package:adobe_xd/specific_rect_clip.dart';
-
-class iPhoneXXSpage221 extends StatelessWidget {
-  iPhoneXXSpage221({
-    Key key,
-  }) : super(key: key);
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: const Color(0xffffffff),
-      body: Stack(
-        children: <Widget>[
-          Container(
-            width: 375.0,
-            height: 190.0,
-            decoration: BoxDecoration(
-              color: const Color(0x2bff871c),
-            ),
-          ),
-          Transform.translate(
-            offset: Offset(-20.0, -5.0),
-            child:
-                // Adobe XD layer: 'Background' (shape)
-                SvgPicture.string(
-              _svg_a1kxap,
-              allowDrawingOutsideViewBox: true,
-            ),
-          ),
-          Transform.translate(
-            offset: Offset(329.67, 15.39),
-            child:
-                // Adobe XD layer: 'Battery' (group)
-                Stack(
-              children: <Widget>[
-                // Adobe XD layer: 'Shape' (shape)
-                SvgPicture.string(
-                  _svg_xx0ljg,
-                  allowDrawingOutsideViewBox: true,
-                ),
-              ],
-            ),
-          ),
-          Transform.translate(
-            offset: Offset(309.37, 15.38),
-            child:
-                // Adobe XD layer: 'Shape' (shape)
-                SvgPicture.string(
-              _svg_8txzf6,
-              allowDrawingOutsideViewBox: true,
-            ),
-          ),
-          Transform.translate(
-            offset: Offset(287.34, 15.73),
-            child:
-                // Adobe XD layer: 'Cellular Connection' (group)
-                Stack(
-              children: <Widget>[
-                // Adobe XD layer: 'Shape' (shape)
-                SvgPicture.string(
-                  _svg_31ra30,
-                  allowDrawingOutsideViewBox: true,
-                ),
-              ],
-            ),
-          ),
-          Transform.translate(
-            offset: Offset(21.0, 11.0),
-            child:
-                // Adobe XD layer: 'Bars/_/Time Black' (group)
-                Stack(
-              children: <Widget>[
-                // Adobe XD layer: 'Background' (shape)
-                SvgPicture.string(
-                  _svg_socz7d,
-                  allowDrawingOutsideViewBox: true,
-                ),
-                Transform.translate(
-                  offset: Offset(0.0, 2.0),
-                  child:
-                      // Adobe XD layer: '↳ Time' (text)
-                      SizedBox(
-                    width: 30.0,
-                    height: 16.0,
-                    child: Text(
-                      '8:35',
-                      style: TextStyle(
-                        fontFamily: 'Helvetica',
-                        fontSize: 14,
-                        color: const Color(0xff343638),
-                        letterSpacing: 0.003920000016689301,
-                      ),
-                      textAlign: TextAlign.left,
-                    ),
-                  ),
-                ),
-              ],
-            ),
-          ),
-          Transform.translate(
-            offset: Offset(0.0, 761.0),
-            child: Container(
-              width: 375.0,
-              height: 51.0,
-              decoration: BoxDecoration(
-                color: const Color(0xfff2f2f2),
-              ),
-            ),
-          ),
-          Transform.translate(
-            offset: Offset(0.0, 1.0),
-            child: Stack(
-              children: <Widget>[
-                Transform.translate(
-                  offset: Offset(180.0, 777.0),
-                  child: Container(
-                    width: 17.0,
-                    height: 17.0,
-                    decoration: BoxDecoration(
-                      borderRadius: BorderRadius.circular(7.0),
-                      border: Border.all(
-                          width: 2.0, color: const Color(0xff707070)),
-                    ),
-                  ),
-                ),
-                Transform.translate(
-                  offset: Offset(74.5, 778.5),
-                  child: SvgPicture.string(
-                    _svg_nbwld4,
-                    allowDrawingOutsideViewBox: true,
-                  ),
-                ),
-                Transform.translate(
-                  offset: Offset(-2149.0, 0.0),
-                  child: Stack(
-                    children: <Widget>[
-                      Transform.translate(
-                        offset: Offset(2442.09, 779.27),
-                        child: SvgPicture.string(
-                          _svg_c9w201,
-                          allowDrawingOutsideViewBox: true,
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
-              ],
-            ),
-          ),
-          Transform.translate(
-            offset: Offset(23.73, 9.72),
-            child: Stack(
-              children: <Widget>[
-                Transform(
-                  transform: Matrix4(0.0, 1.0, 0.0, 0.0, -1.0, 0.0, 0.0, 0.0,
-                      0.0, 0.0, 1.0, 0.0, 323.64, 67.23, 0.0, 1.0),
-                  child:
-                      // Adobe XD layer: 'next / previous' (group)
-                      Stack(
-                    children: <Widget>[
-                      Transform.translate(
-                        offset: Offset(15.6, 0.0),
-                        child:
-                            // Adobe XD layer: 'arrow' (group)
-                            Stack(
-                          children: <Widget>[
-                            Transform.translate(
-                              offset: Offset(0.0, 0.0),
-                              child:
-                                  // Adobe XD layer: 'arrow' (shape)
-                                  SvgPicture.string(
-                                _svg_6w15ai,
-                                allowDrawingOutsideViewBox: true,
-                              ),
-                            ),
-                          ],
-                        ),
-                      ),
-                      Transform.translate(
-                        offset: Offset(0.0, 5.42),
-                        child:
-                            // Adobe XD layer: 'arrow back' (group)
-                            Stack(
-                          children: <Widget>[
-                            // Adobe XD layer: 'arrow back' (shape)
-                            SvgPicture.string(
-                              _svg_m1usag,
-                              allowDrawingOutsideViewBox: true,
-                            ),
-                          ],
-                        ),
-                      ),
-                      Transform.translate(
-                        offset: Offset(1.82, 5.49),
-                        child: SvgPicture.string(
-                          _svg_t4lnty,
-                          allowDrawingOutsideViewBox: true,
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
-              ],
-            ),
-          ),
-          Transform.translate(
-            offset: Offset(52.73, 45.0),
-            child: Stack(
-              children: <Widget>[
-                Container(
-                  width: 270.0,
-                  height: 39.0,
-                  decoration: BoxDecoration(
-                    borderRadius: BorderRadius.circular(20.0),
-                    color: const Color(0xffffffff),
-                    border:
-                        Border.all(width: 1.0, color: const Color(0xffff9100)),
-                  ),
-                ),
-                Transform.translate(
-                  offset: Offset(51.0, 11.59),
-                  child: Text(
-                    '내 위치: 부산광역시 금정구 장전동...',
-                    style: TextStyle(
-                      fontFamily: 'BM JUA_TTF',
-                      fontSize: 14,
-                      color: const Color(0xfff0ad74),
-                      height: 1.2142857142857142,
-                    ),
-                    textAlign: TextAlign.left,
-                  ),
-                ),
-                Transform.translate(
-                  offset: Offset(19.0, 11.59),
-                  child: Text(
-                    '출발',
-                    style: TextStyle(
-                      fontFamily: 'BM JUA_TTF',
-                      fontSize: 14,
-                      color: const Color(0x80f0ad74),
-                      height: 1.2142857142857142,
-                    ),
-                    textAlign: TextAlign.left,
-                  ),
-                ),
-              ],
-            ),
-          ),
-          Transform.translate(
-            offset: Offset(52.73, 92.0),
-            child: Stack(
-              children: <Widget>[
-                Container(
-                  width: 270.0,
-                  height: 39.0,
-                  decoration: BoxDecoration(
-                    borderRadius: BorderRadius.circular(20.0),
-                    color: const Color(0xffffffff),
-                    border:
-                        Border.all(width: 1.0, color: const Color(0xffff9100)),
-                  ),
-                ),
-                Transform.translate(
-                  offset: Offset(51.0, 11.59),
-                  child: Text(
-                    '도착지: 장소 또는 주소 입력',
-                    style: TextStyle(
-                      fontFamily: 'BM JUA_TTF',
-                      fontSize: 14,
-                      color: const Color(0xfff0ad74),
-                      height: 1.2142857142857142,
-                    ),
-                    textAlign: TextAlign.left,
-                  ),
-                ),
-                Transform.translate(
-                  offset: Offset(19.0, 11.59),
-                  child: Text(
-                    '도착',
-                    style: TextStyle(
-                      fontFamily: 'BM JUA_TTF',
-                      fontSize: 14,
-                      color: const Color(0x80f0ad74),
-                      height: 1.2142857142857142,
-                    ),
-                    textAlign: TextAlign.left,
-                  ),
-                ),
-              ],
-            ),
-          ),
-          Transform.translate(
-            offset: Offset(55.0, 218.1),
-            child: Stack(
-              children: <Widget>[
-                Transform.translate(
-                  offset: Offset(276.87, -71.0),
-                  child:
-                      // Adobe XD layer: 'star' (shape)
-                      SvgPicture.string(
-                    _svg_f1fn4l,
-                    allowDrawingOutsideViewBox: true,
-                  ),
-                ),
-              ],
-            ),
-          ),
-          Stack(
-            children: <Widget>[
-              Transform.translate(
-                offset: Offset(26.37, 50.14),
-                child: SvgPicture.string(
-                  _svg_f6aidi,
-                  allowDrawingOutsideViewBox: true,
-                ),
-              ),
-            ],
-          ),
-          Transform.translate(
-            offset: Offset(21.0, 201.83),
-            child: Text(
-              '최근 검색',
-              style: TextStyle(
-                fontFamily: 'BM JUA_TTF',
-                fontSize: 8,
-                color: const Color(0xff636363),
-                height: 1.375,
-              ),
-              textAlign: TextAlign.left,
-            ),
-          ),
-          Transform.translate(
-            offset: Offset(35.0, 222.0),
-            child: SpecificRectClip(
-              rect: Rect.fromLTWH(0, 0, 340, 521.3323974609375),
-              child: UnconstrainedBox(
-                alignment: Alignment.topLeft,
-                child: Container(
-                  width: 340,
-                  height: 511,
-                  child: GridView.count(
-                    primary: false,
-                    padding: EdgeInsets.all(0),
-                    mainAxisSpacing: 20,
-                    crossAxisSpacing: 20,
-                    crossAxisCount: 1,
-                    childAspectRatio: 8.7179,
-                    children: [
-                      {
-                        'text': '부산광역시 금정구 장전3동 정전로12번길',
-                      },
-                      {
-                        'text': '부산광역시 금정구 금강로335번길 90 (장전동)',
-                      },
-                      {
-                        'text': '부산광역시 금정구 부산대학로 53 1층 (장전동)',
-                      },
-                      {
-                        'text': '부산광역시 금정구 장전3동 정전로12번길',
-                      },
-                      {
-                        'text': '부산광역시 금정구 장전3동 정전로12번길',
-                      },
-                      {
-                        'text': '부산광역시 금정구 장전3동 정전로12번길',
-                      },
-                      {
-                        'text': '부산광역시 금정구 장전3동 정전로12번길',
-                      },
-                      {
-                        'text': '부산광역시 금정구 장전3동 정전로12번길',
-                      },
-                      {
-                        'text': '부산광역시 금정구 장전3동 정전로12번길',
-                      },
-                    ].map((map) {
-                      final text = map['text'];
-                      return Transform.translate(
-                        offset: Offset(-35.0, -203.0),
-                        child: Stack(
-                          children: <Widget>[
-                            Transform.translate(
-                              offset: Offset(72.79, 240.83),
-                              child: SvgPicture.string(
-                                _svg_zbs4ll,
-                                allowDrawingOutsideViewBox: true,
-                              ),
-                            ),
-                            Transform.translate(
-                              offset: Offset(35.0, 203.0),
-                              child: Stack(
-                                children: <Widget>[
-                                  Transform.translate(
-                                    offset: Offset(40.0, 4.83),
-                                    child: Text(
-                                      text,
-                                      style: TextStyle(
-                                        fontFamily: 'BM JUA_TTF',
-                                        fontSize: 14,
-                                        color: const Color(0xff000000),
-                                        height: 0.7857142857142857,
-                                      ),
-                                      textAlign: TextAlign.left,
-                                    ),
-                                  ),
-                                  Transform.translate(
-                                    offset: Offset(40.0, 19.83),
-                                    child: Text(
-                                      text,
-                                      style: TextStyle(
-                                        fontFamily: 'BM JUA_TTF',
-                                        fontSize: 10,
-                                        color: const Color(0x91000000),
-                                        height: 1.1,
-                                      ),
-                                      textAlign: TextAlign.left,
-                                    ),
-                                  ),
-                                  Transform.translate(
-                                    offset: Offset(0.0, 2.0),
-                                    child: Stack(
-                                      children: <Widget>[
-                                        Container(
-                                          width: 28.0,
-                                          height: 28.0,
-                                          decoration: BoxDecoration(
-                                            borderRadius: BorderRadius.all(
-                                                Radius.elliptical(14.0, 14.0)),
-                                            color: const Color(0x80ff5807),
-                                          ),
-                                        ),
-                                        Transform.translate(
-                                          offset: Offset(4.67, 4.45),
-                                          child:
-                                              // Adobe XD layer: 'time' (group)
-                                              Stack(
-                                            children: <Widget>[
-                                              SvgPicture.string(
-                                                _svg_g3jfrp,
-                                                allowDrawingOutsideViewBox:
-                                                    true,
-                                              ),
-                                            ],
-                                          ),
-                                        ),
-                                      ],
-                                    ),
-                                  ),
-                                ],
-                              ),
-                            ),
-                          ],
-                        ),
-                      );
-                    }).toList(),
-                  ),
-                ),
-              ),
-            ),
-          ),
-          Transform.translate(
-            offset: Offset(-58.0, 0.0),
-            child: Stack(
-              children: <Widget>[
-                Transform.translate(
-                  offset: Offset(113.64, 156.01),
-                  child: Text(
-                    '집',
-                    style: TextStyle(
-                      fontFamily: 'BM JUA_TTF',
-                      fontSize: 12,
-                      color: const Color(0xff080808),
-                      height: 0.9166666666666666,
-                    ),
-                    textAlign: TextAlign.left,
-                  ),
-                ),
-                Transform.translate(
-                  offset: Offset(1.0, 0.0),
-                  child: Stack(
-                    children: <Widget>[
-                      Transform.translate(
-                        offset: Offset(78.0, 148.0),
-                        child: Container(
-                          width: 25.0,
-                          height: 25.0,
-                          decoration: BoxDecoration(
-                            borderRadius:
-                                BorderRadius.all(Radius.elliptical(12.5, 12.5)),
-                            color: const Color(0xffff871c),
-                          ),
-                        ),
-                      ),
-                      Transform.translate(
-                        offset: Offset(84.55, 153.64),
-                        child: SvgPicture.string(
-                          _svg_8512wh,
-                          allowDrawingOutsideViewBox: true,
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
-              ],
-            ),
-          ),
-          Transform.translate(
-            offset: Offset(75.0, 0.0),
-            child: Stack(
-              children: <Widget>[
-                Transform.translate(
-                  offset: Offset(113.64, 156.01),
-                  child: Text(
-                    '도서관',
-                    style: TextStyle(
-                      fontFamily: 'BM JUA_TTF',
-                      fontSize: 12,
-                      color: const Color(0xff080808),
-                      height: 0.9166666666666666,
-                    ),
-                    textAlign: TextAlign.left,
-                  ),
-                ),
-                Transform.translate(
-                  offset: Offset(1.0, 0.0),
-                  child: Stack(
-                    children: <Widget>[
-                      Stack(
-                        children: <Widget>[
-                          Transform.translate(
-                            offset: Offset(78.0, 148.0),
-                            child: Container(
-                              width: 25.0,
-                              height: 25.0,
-                              decoration: BoxDecoration(
-                                borderRadius: BorderRadius.all(
-                                    Radius.elliptical(12.5, 12.5)),
-                                color: const Color(0xffff871c),
-                              ),
-                            ),
-                          ),
-                          Transform.translate(
-                            offset: Offset(85.77, 153.59),
-                            child: SvgPicture.string(
-                              _svg_9nrqm3,
-                              allowDrawingOutsideViewBox: true,
-                            ),
-                          ),
-                        ],
-                      ),
-                    ],
-                  ),
-                ),
-              ],
-            ),
-          ),
-          Transform.translate(
-            offset: Offset(-56.0, 0.0),
-            child: Stack(
-              children: <Widget>[
-                Transform.translate(
-                  offset: Offset(172.16, 156.01),
-                  child: Text(
-                    '학교',
-                    style: TextStyle(
-                      fontFamily: 'BM JUA_TTF',
-                      fontSize: 12,
-                      color: const Color(0xff080808),
-                      height: 0.9166666666666666,
-                    ),
-                    textAlign: TextAlign.left,
-                  ),
-                ),
-                Stack(
-                  children: <Widget>[
-                    Transform.translate(
-                      offset: Offset(138.0, 148.0),
-                      child: Container(
-                        width: 25.0,
-                        height: 25.0,
-                        decoration: BoxDecoration(
-                          borderRadius:
-                              BorderRadius.all(Radius.elliptical(12.5, 12.5)),
-                          color: const Color(0xffff871c),
-                        ),
-                      ),
-                    ),
-                    Transform.translate(
-                      offset: Offset(144.44, 153.37),
-                      child: SvgPicture.string(
-                        _svg_gow1z9,
-                        allowDrawingOutsideViewBox: true,
-                      ),
-                    ),
-                  ],
-                ),
-              ],
-            ),
-          ),
-          Transform.translate(
-            offset: Offset(97.0, 0.0),
-            child: Stack(
-              children: <Widget>[
-                Transform.translate(
-                  offset: Offset(172.16, 156.01),
-                  child: Text(
-                    '수정',
-                    style: TextStyle(
-                      fontFamily: 'BM JUA_TTF',
-                      fontSize: 12,
-                      color: const Color(0xff080808),
-                      height: 0.9166666666666666,
-                    ),
-                    textAlign: TextAlign.left,
-                  ),
-                ),
-                Stack(
-                  children: <Widget>[
-                    Transform.translate(
-                      offset: Offset(138.0, 148.0),
-                      child: Container(
-                        width: 25.0,
-                        height: 25.0,
-                        decoration: BoxDecoration(
-                          borderRadius:
-                              BorderRadius.all(Radius.elliptical(12.5, 12.5)),
-                          color: const Color(0xffff871c),
-                        ),
-                      ),
-                    ),
-                    Transform.translate(
-                      offset: Offset(143.65, 151.0),
-                      child: SvgPicture.string(
-                        _svg_4lxdll,
-                        allowDrawingOutsideViewBox: true,
-                      ),
-                    ),
-                    Transform(
-                      transform: Matrix4(
-                          0.325568,
-                          0.945519,
-                          0.0,
-                          0.0,
-                          -0.945519,
-                          0.325568,
-                          0.0,
-                          0.0,
-                          0.0,
-                          0.0,
-                          1.0,
-                          0.0,
-                          154.52,
-                          160.66,
-                          0.0,
-                          1.0),
-                      child: Container(
-                        width: 4.0,
-                        height: 5.0,
-                        decoration: BoxDecoration(
-                          borderRadius: BorderRadius.circular(1.0),
-                          color: const Color(0xffffffff),
-                        ),
-                      ),
-                    ),
-                  ],
-                ),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-const String _svg_xx0ljg =
-    '<svg viewBox="0.0 0.0 24.3 11.3" ><path transform="translate(-336.0, 26.67)" d="M 338.6666870117188 -26.66666984558105 L 355.3333129882812 -26.66666984558105 L 355.3333129882812 -26.66666984558105 C 356.8060913085938 -26.66666984558105 358 -25.47275924682617 358 -24 L 358 -18 L 358 -18 C 358 -16.52724075317383 356.8060913085938 -15.33333015441895 355.3333129882812 -15.33333015441895 L 338.6666870117188 -15.33333015441895 L 338.6666870117188 -15.33333015441895 C 337.1939086914062 -15.33333015441895 336 -16.52724075317383 336 -18 L 336 -24 L 336 -24 C 336 -25.47275924682617 337.1939086914062 -26.66666984558105 338.6666870117188 -26.66666984558105 Z" fill="none" fill-opacity="0.35" stroke="#343638" stroke-width="1" stroke-opacity="0.35" stroke-miterlimit="4" stroke-linecap="butt" /><path transform="translate(-336.0, 26.67)" d="M 359 -23 L 359 -19 C 359.8046875 -19.33877944946289 360.3280029296875 -20.12686920166016 360.3280029296875 -21 C 360.3280029296875 -21.87313079833984 359.8046875 -22.66122055053711 359 -23" fill="#343638" stroke="none" stroke-width="1" stroke-miterlimit="4" stroke-linecap="butt" /><path transform="translate(-336.0, 26.67)" d="M 339.3333129882812 -24.66666984558105 L 354.6666870117188 -24.66666984558105 L 354.6666870117188 -24.66666984558105 C 355.4030151367188 -24.66666984558105 356 -24.06970977783203 356 -23.33333015441895 L 356 -18.66666984558105 L 356 -18.66666984558105 C 356 -17.93029022216797 355.4030151367188 -17.33333015441895 354.6666870117188 -17.33333015441895 L 339.3333129882812 -17.33333015441895 L 339.3333129882812 -17.33333015441895 C 338.5969848632812 -17.33333015441895 338 -17.93029022216797 338 -18.66666984558105 L 338 -23.33333015441895 L 338 -23.33333015441895 C 338 -24.06970977783203 338.5969848632812 -24.66666984558105 339.3333129882812 -24.66666984558105 Z" fill="#343638" stroke="none" stroke-width="1" stroke-miterlimit="4" stroke-linecap="butt" /></svg>';
-const String _svg_a1kxap =
-    '<svg viewBox="-20.0 -5.0 375.0 46.0" ><path transform="translate(-20.0, 39.0)" d="M 0 -44 L 375 -44 L 375 2 L 0 2 L 0 -44 Z" fill="none" stroke="none" stroke-width="1" stroke-miterlimit="4" stroke-linecap="butt" /></svg>';
-const String _svg_31ra30 =
-    '<svg viewBox="0.0 0.0 17.0 10.7" ><path transform="translate(-293.67, 26.33)" d="M 294.6666870117188 -19.66666984558105 L 295.6666870117188 -19.66666984558105 C 296.218994140625 -19.66666984558105 296.6666870117188 -19.21895027160645 296.6666870117188 -18.66666984558105 L 296.6666870117188 -16.66666984558105 C 296.6666870117188 -16.1143798828125 296.218994140625 -15.66666984558105 295.6666870117188 -15.66666984558105 L 294.6666870117188 -15.66666984558105 C 294.1144104003906 -15.66666984558105 293.6666870117188 -16.1143798828125 293.6666870117188 -16.66666984558105 L 293.6666870117188 -18.66666984558105 C 293.6666870117188 -19.21895027160645 294.1144104003906 -19.66666984558105 294.6666870117188 -19.66666984558105 L 294.6666870117188 -19.66666984558105 Z M 299.3333129882812 -21.66666984558105 L 300.3333129882812 -21.66666984558105 C 300.8855895996094 -21.66666984558105 301.3333129882812 -21.21895027160645 301.3333129882812 -20.66666984558105 L 301.3333129882812 -16.66666984558105 C 301.3333129882812 -16.1143798828125 300.8855895996094 -15.66666984558105 300.3333129882812 -15.66666984558105 L 299.3333129882812 -15.66666984558105 C 298.781005859375 -15.66666984558105 298.3333129882812 -16.1143798828125 298.3333129882812 -16.66666984558105 L 298.3333129882812 -20.66666984558105 C 298.3333129882812 -21.21895027160645 298.781005859375 -21.66666984558105 299.3333129882812 -21.66666984558105 Z M 304 -24 L 305 -24 C 305.5523071289062 -24 306 -23.55228042602539 306 -23 L 306 -16.66666984558105 C 306 -16.1143798828125 305.5523071289062 -15.66666984558105 305 -15.66666984558105 L 304 -15.66666984558105 C 303.4476928710938 -15.66666984558105 303 -16.1143798828125 303 -16.66666984558105 L 303 -23 C 303 -23.55228042602539 303.4476928710938 -24 304 -24 Z M 308.6666870117188 -26.33333015441895 L 309.6666870117188 -26.33333015441895 C 310.218994140625 -26.33333015441895 310.6666870117188 -25.8856201171875 310.6666870117188 -25.33333015441895 L 310.6666870117188 -16.66666984558105 C 310.6666870117188 -16.1143798828125 310.218994140625 -15.66666984558105 309.6666870117188 -15.66666984558105 L 308.6666870117188 -15.66666984558105 C 308.1144104003906 -15.66666984558105 307.6666870117188 -16.1143798828125 307.6666870117188 -16.66666984558105 L 307.6666870117188 -25.33333015441895 C 307.6666870117188 -25.8856201171875 308.1144104003906 -26.33333015441895 308.6666870117188 -26.33333015441895 L 308.6666870117188 -26.33333015441895 Z" fill="#343638" stroke="none" stroke-width="1" stroke-miterlimit="4" stroke-linecap="butt" /></svg>';
-const String _svg_8txzf6 =
-    '<svg viewBox="309.4 15.4 15.3 11.0" ><path transform="translate(-6.33, 42.05)" d="M 323.3302917480469 -24.39200019836426 C 325.5462036132812 -24.39189910888672 327.6773986816406 -23.54047012329102 329.2832946777344 -22.01366996765137 C 329.404296875 -21.89579010009766 329.5975952148438 -21.89727973937988 329.7167053222656 -22.01700019836426 L 330.8727111816406 -23.18367004394531 C 330.9330139160156 -23.2443904876709 330.9666137695312 -23.32663917541504 330.9660949707031 -23.4122200012207 C 330.9656066894531 -23.49780082702637 330.9309997558594 -23.57965087890625 330.8699951171875 -23.63966941833496 C 326.6549072265625 -27.6792106628418 320.0050964355469 -27.6792106628418 315.7900085449219 -23.63966941833496 C 315.7289123535156 -23.5797004699707 315.6943054199219 -23.49786949157715 315.6936950683594 -23.41229057312012 C 315.6932067871094 -23.32670974731445 315.7267150878906 -23.24443054199219 315.7869873046875 -23.18367004394531 L 316.9432983398438 -22.01700019836426 C 317.0624084472656 -21.8971004486084 317.2557983398438 -21.89561080932617 317.376708984375 -22.01366996765137 C 318.9827880859375 -23.54056930541992 321.1141967773438 -24.39200973510742 323.3302917480469 -24.39200019836426 L 323.3302917480469 -24.39200019836426 Z M 323.3302917480469 -20.5963306427002 C 324.5477905273438 -20.59641075134277 325.7218933105469 -20.14388084411621 326.6242980957031 -19.32666969299316 C 326.7463989257812 -19.21068954467773 326.9386901855469 -19.21319961547852 327.0577087402344 -19.33233070373535 L 328.2123107910156 -20.49900054931641 C 328.2731018066406 -20.56019020080566 328.306884765625 -20.64320945739746 328.3059997558594 -20.7294807434082 C 328.3051147460938 -20.81574058532715 328.2697143554688 -20.89805030822754 328.2077026367188 -20.95800018310547 C 325.4595031738281 -23.51437950134277 321.2034912109375 -23.51437950134277 318.4552917480469 -20.95800018310547 C 318.393310546875 -20.89805030822754 318.3577880859375 -20.81570053100586 318.3569946289062 -20.72941017150879 C 318.356201171875 -20.64311981201172 318.3901062011719 -20.56011009216309 318.4509887695312 -20.49900054931641 L 319.6052856445312 -19.33233070373535 C 319.7243041992188 -19.21319961547852 319.9165954589844 -19.21068954467773 320.0386962890625 -19.32666969299316 C 320.9404907226562 -20.14333915710449 322.1135864257812 -20.5958309173584 323.3302917480469 -20.5963306427002 L 323.3302917480469 -20.5963306427002 Z M 325.5492858886719 -17.8120002746582 C 325.6111145019531 -17.87261009216309 325.6451110839844 -17.95601081848145 325.643310546875 -18.04250907897949 C 325.6416015625 -18.12902069091797 325.6041870117188 -18.21096992492676 325.5400085449219 -18.26899909973145 C 324.264404296875 -19.34787940979004 322.3962097167969 -19.34787940979004 321.1206970214844 -18.26899909973145 C 321.056396484375 -18.21100997924805 321.0190124511719 -18.12908935546875 321.0172119140625 -18.04258918762207 C 321.0152893066406 -17.95607948303223 321.0492858886719 -17.87265014648438 321.1109924316406 -17.8120002746582 L 323.1087036132812 -15.79633045196533 C 323.1672058105469 -15.73709011077881 323.2470092773438 -15.70376014709473 323.3302917480469 -15.70376014709473 C 323.4136047363281 -15.70376014709473 323.493408203125 -15.73709011077881 323.552001953125 -15.79633045196533 L 325.5492858886719 -17.8120002746582 Z" fill="#343638" stroke="none" stroke-width="1" stroke-miterlimit="4" stroke-linecap="butt" /></svg>';
-const String _svg_socz7d =
-    '<svg viewBox="0.0 0.0 30.2 16.0" ><path transform="translate(-21.0, 35.0)" d="M 21 -19 L 51.21133041381836 -19 L 51.21133041381836 -35 L 21 -35 L 21 -19 Z" fill="none" stroke="none" stroke-width="1" stroke-miterlimit="4" stroke-linecap="butt" /></svg>';
-const String _svg_c9w201 =
-    '<svg viewBox="2442.1 779.3 6.8 13.8" ><path transform="matrix(0.707107, 0.707107, -0.707107, 0.707107, 2448.91, 779.27)" d="M 0 0 L 0 9.63812255859375" fill="none" stroke="#707070" stroke-width="2" stroke-miterlimit="4" stroke-linecap="round" /><path transform="matrix(-0.707107, 0.707107, -0.707107, -0.707107, 2448.91, 793.09)" d="M 0 0 L 0 9.63812255859375" fill="none" stroke="#707070" stroke-width="2" stroke-miterlimit="4" stroke-linecap="round" /></svg>';
-const String _svg_nbwld4 =
-    '<svg viewBox="74.5 778.5 12.0 15.0" ><path transform="translate(74.5, 778.5)" d="M 0 0 L 0 15" fill="none" stroke="#707070" stroke-width="2" stroke-miterlimit="4" stroke-linecap="round" /><path transform="translate(80.5, 778.5)" d="M 0 0 L 0 15" fill="none" stroke="#707070" stroke-width="2" stroke-miterlimit="4" stroke-linecap="round" /><path transform="translate(86.5, 778.5)" d="M 0 0 L 0 15" fill="none" stroke="#707070" stroke-width="2" stroke-miterlimit="4" stroke-linecap="round" /></svg>';
-const String _svg_6w15ai =
-    '<svg viewBox="0.0 0.0 6.5 10.8" ><path transform="matrix(0.0, -1.0, 1.0, 0.0, 0.0, 10.85)" d="M 5.422946929931641 6.507537364959717 C 5.097570419311523 6.507537364959717 4.88065242767334 6.399077892303467 4.663734912872314 6.182159423828125 L 0.3253768682479858 1.843802213668823 C -0.108458936214447 1.409966230392456 -0.108458936214447 0.7592126131057739 0.3253768682479858 0.3253768682479858 C 0.7592126131057739 -0.1084589064121246 1.409966230392456 -0.108458936214447 1.843802213668823 0.3253768682479858 L 5.422946929931641 3.904522180557251 L 9.002093315124512 0.3253769874572754 C 9.435928344726562 -0.1084588095545769 10.08668231964111 -0.1084588095545769 10.52051830291748 0.3253769874572754 C 10.95435333251953 0.7592127323150635 10.95435333251953 1.409966588020325 10.52051830291748 1.843802213668823 L 6.182159423828125 6.182159423828125 C 5.965242385864258 6.399077892303467 5.748324394226074 6.507537364959717 5.422946929931641 6.507537364959717 Z" fill="#ffb300" stroke="none" stroke-width="1" stroke-miterlimit="4" stroke-linecap="butt" /></svg>';
-const String _svg_m1usag =
-    '<svg viewBox="0.0 0.0 6.5 10.8" ><path transform="matrix(0.0, 1.0, -1.0, 0.0, 6.51, 0.0)" d="M 5.422946929931641 6.507537364959717 C 5.097570419311523 6.507537364959717 4.88065242767334 6.399077892303467 4.663734912872314 6.182159423828125 L 0.3253768682479858 1.843802213668823 C -0.108458936214447 1.409966230392456 -0.108458936214447 0.7592126131057739 0.3253768682479858 0.3253768682479858 C 0.7592126131057739 -0.1084589064121246 1.409966230392456 -0.108458936214447 1.843802213668823 0.3253768682479858 L 5.422946929931641 3.904522180557251 L 9.002093315124512 0.3253769874572754 C 9.435928344726562 -0.1084588095545769 10.08668231964111 -0.1084588095545769 10.52051830291748 0.3253769874572754 C 10.95435333251953 0.7592127323150635 10.95435333251953 1.409966588020325 10.52051830291748 1.843802213668823 L 6.182159423828125 6.182159423828125 C 5.965242385864258 6.399077892303467 5.748324394226074 6.507537364959717 5.422946929931641 6.507537364959717 Z" fill="#ffb300" stroke="none" stroke-width="1" stroke-miterlimit="4" stroke-linecap="butt" /></svg>';
-const String _svg_t4lnty =
-    '<svg viewBox="1.8 5.5 18.5 5.4" ><path transform="translate(1.82, 10.86)" d="M 8.830299377441406 0 L 0 0" fill="none" stroke="#ffb300" stroke-width="3" stroke-miterlimit="4" stroke-linecap="round" /><path transform="translate(11.46, 5.49)" d="M 8.830299377441406 0 L 0 0" fill="none" stroke="#ffb300" stroke-width="3" stroke-miterlimit="4" stroke-linecap="round" /></svg>';
-const String _svg_f1fn4l =
-    '<svg viewBox="276.9 -71.0 22.1 21.2" ><path transform="translate(276.87, -71.49)" d="M 22.07576370239258 8.506759643554688 C 21.93087387084961 8.05858325958252 21.53335762023926 7.740263938903809 21.06305885314941 7.697882175445557 L 14.67477989196777 7.117817401885986 L 12.1486759185791 1.205204129219055 C 11.96241474151611 0.771889328956604 11.53821754455566 0.4913979470729828 11.0669059753418 0.4913979470729828 C 10.59559535980225 0.4913979470729828 10.1713981628418 0.7718884348869324 9.985136985778809 1.206219911575317 L 7.459032535552979 7.117795944213867 L 1.069718599319458 7.697860240936279 C 0.6002630591392517 7.741259098052979 0.2037586718797684 8.058564186096191 0.05802920833230019 8.50673770904541 C -0.08770414441823959 8.954916000366211 0.04688385501503944 9.446492195129395 0.402014285326004 9.756367683410645 L 5.230799674987793 13.99125003814697 L 3.806900501251221 20.26354217529297 C 3.70270848274231 20.72472381591797 3.881709098815918 21.20143890380859 4.264364242553711 21.47804832458496 C 4.470046043395996 21.62665176391602 4.71068286895752 21.70230484008789 4.953348159790039 21.70230484008789 C 5.162575721740723 21.70230484008789 5.370115756988525 21.64590263366699 5.556376457214355 21.53445053100586 L 11.06690406799316 18.24100685119629 L 16.57540130615234 21.53445053100586 C 16.97849082946777 21.77694511413574 17.48661422729492 21.75482559204102 17.86842727661133 21.47804832458496 C 18.25125312805176 21.20059776306152 18.43008232116699 20.72371292114258 18.32589340209961 20.26354217529297 L 16.90199279785156 13.99125003814697 L 21.73077583312988 9.757209777832031 C 22.08590888977051 9.446491241455078 22.22150993347168 8.955759048461914 22.07577705383301 8.50673770904541 Z M 22.07576370239258 8.506759643554688" fill="#ff871c" stroke="none" stroke-width="1" stroke-miterlimit="4" stroke-linecap="butt" /></svg>';
-const String _svg_f6aidi =
-    '<svg viewBox="26.4 50.1 14.7 14.7" ><path transform="matrix(0.707107, 0.707107, -0.707107, 0.707107, 28.36, 50.14)" d="M 1 0 L 17 0 C 17.55228424072266 0 18 0.2522549033164978 18 0.5634271502494812 L 18 2.253708600997925 C 18 2.564880847930908 17.55228424072266 2.817135810852051 17 2.817135810852051 C 17 2.817135810852051 12.8143196105957 2.817135810852051 8.814319610595703 2.817135810852051 C 4.814319610595703 2.817135810852051 1 2.817135810852051 1 2.817135810852051 C 0.4477152228355408 2.817135810852051 0 2.564880847930908 0 2.253708600997925 L 0 0.5634271502494812 C 0 0.2522549033164978 0.4477152228355408 0 1 0 Z" fill="#ffb300" stroke="none" stroke-width="1" stroke-miterlimit="4" stroke-linecap="butt" /><path transform="matrix(0.707107, -0.707107, 0.707107, 0.707107, 26.37, 62.87)" d="M 1 0 L 17 0 C 17.55228424072266 0 18 0.2522549033164978 18 0.5634271502494812 L 18 2.253708600997925 C 18 2.564880847930908 17.55228424072266 2.817135810852051 17 2.817135810852051 C 17 2.817135810852051 12.8143196105957 2.817135810852051 8.814319610595703 2.817135810852051 C 4.814319610595703 2.817135810852051 1 2.817135810852051 1 2.817135810852051 C 0.4477152228355408 2.817135810852051 0 2.564880847930908 0 2.253708600997925 L 0 0.5634271502494812 C 0 0.2522549033164978 0.4477152228355408 0 1 0 Z" fill="#ffb300" stroke="none" stroke-width="1" stroke-miterlimit="4" stroke-linecap="butt" /></svg>';
-const String _svg_g3jfrp =
-    '<svg viewBox="0.0 0.0 18.9 18.9" ><path  d="M 9.440127372741699 0 C 4.24805736541748 0 0 4.24805736541748 0 9.440127372741699 C 0 14.6321964263916 4.24805736541748 18.8802547454834 9.440127372741699 18.8802547454834 C 14.6321964263916 18.8802547454834 18.8802547454834 14.6321964263916 18.8802547454834 9.440127372741699 C 18.8802547454834 4.24805736541748 14.6321964263916 0 9.440127372741699 0 Z M 9.440127372741699 16.52022171020508 C 5.546074390411377 16.52022171020508 2.360031843185425 13.33417987823486 2.360031843185425 9.440127372741699 C 2.360031843185425 5.546074390411377 5.546074390411377 2.360031843185425 9.440127372741699 2.360031843185425 C 13.33417987823486 2.360031843185425 16.52022171020508 5.546074390411377 16.52022171020508 9.440127372741699 C 16.52022171020508 13.33417987823486 13.33417987823486 16.52022171020508 9.440127372741699 16.52022171020508 Z M 10.62014293670654 8.968120574951172 L 13.80618667602539 12.15416431427002 L 12.15416431427002 13.80618667602539 L 8.260110855102539 9.91213321685791 L 8.260110855102539 3.540047645568848 L 10.62014293670654 3.540047645568848 L 10.62014293670654 8.968120574951172 Z" fill="#fff69c" stroke="none" stroke-width="1" stroke-miterlimit="4" stroke-linecap="butt" /></svg>';
-const String _svg_zbs4ll =
-    '<svg viewBox="72.8 240.8 302.2 1.0" ><path transform="translate(72.79, 240.83)" d="M 0 0 L 302.2099609375 0" fill="none" fill-opacity="0.3" stroke="#ff5807" stroke-width="1" stroke-opacity="0.3" stroke-miterlimit="4" stroke-linecap="butt" /></svg>';
-const String _svg_8512wh =
-    '<svg viewBox="84.5 153.6 12.2 13.0" ><path transform="translate(6.44, -97.09)" d="M 89.9969482421875 254.9488677978516 L 84.84983825683594 250.9415740966797 C 84.4918212890625 250.6627655029297 83.97944641113281 250.6627655029297 83.62143707275391 250.9415740966797 L 78.47431945800781 254.9488677978516 C 77.76065063476562 255.5044097900391 78.16979217529297 256.6130065917969 79.08849334716797 256.6130065917969 L 79.08849334716797 256.6130065917969 C 79.13475799560547 256.6130065917969 79.17230224609375 256.64892578125 79.17230224609375 256.693359375 L 79.17230224609375 261.5773620605469 C 79.17230224609375 262.7676086425781 80.17700958251953 263.7324523925781 81.41643524169922 263.7324523925781 L 81.99155426025391 263.7324523925781 C 82.03781127929688 263.7324523925781 82.07529449462891 263.6965026855469 82.07529449462891 263.6520080566406 L 82.07529449462891 259.4858703613281 C 82.07529449462891 258.7253723144531 82.71730041503906 258.1088256835938 83.50931549072266 258.1088256835938 L 84.9620361328125 258.1088256835938 C 85.75397491455078 258.1088256835938 86.39598846435547 258.7253723144531 86.39598846435547 259.4858703613281 L 86.39598846435547 263.6520080566406 C 86.39598846435547 263.6965026855469 86.43352508544922 263.7324523925781 86.47980499267578 263.7324523925781 L 87.05483245849609 263.7324523925781 C 88.29426574707031 263.7324523925781 89.29904174804688 262.7676086425781 89.29904174804688 261.5773620605469 L 89.29904174804688 256.693359375 C 89.29904174804688 256.64892578125 89.33650970458984 256.6130065917969 89.38278198242188 256.6130065917969 L 89.38278198242188 256.6130065917969 C 90.30155181884766 256.6130065917969 90.71070098876953 255.5044097900391 89.9969482421875 254.9488677978516 Z" fill="#ffffff" stroke="none" stroke-width="1" stroke-miterlimit="10" stroke-linecap="butt" /></svg>';
-const String _svg_9nrqm3 =
-    '<svg viewBox="85.8 153.6 9.3 13.0" ><path transform="translate(-571.76, -97.19)" d="M 659.361083984375 260.9382629394531 L 666.4703979492188 260.9382629394531 C 666.6948852539062 260.9382629394531 666.8767700195312 260.7564086914062 666.8767700195312 260.5320129394531 L 666.8767700195312 251.1882629394531 C 666.8767700195312 250.9639587402344 666.6948852539062 250.7820434570312 666.4703979492188 250.7820434570312 L 659.208740234375 250.7820434570312 C 658.2846069335938 250.7820434570312 657.532958984375 251.5337829589844 657.532958984375 252.4578552246094 L 657.532958984375 261.9538879394531 C 657.532958984375 262.9619750976562 658.3529663085938 263.7820434570312 659.361083984375 263.7820434570312 L 666.4703979492188 263.7820434570312 C 666.6948852539062 263.7820434570312 666.8767700195312 263.6001281738281 666.8767700195312 263.3757629394531 C 666.8767700195312 263.1513977050781 666.6948852539062 262.9695739746094 666.4703979492188 262.9695739746094 L 659.361083984375 262.9695739746094 C 658.8010864257812 262.9695739746094 658.345458984375 262.5138549804688 658.345458984375 261.9538879394531 C 658.345458984375 261.3939208984375 658.8010864257812 260.9382629394531 659.361083984375 260.9382629394531 Z M 658.5485229492188 252.2974243164062 C 658.5485229492188 252.0730590820312 658.7305297851562 251.89111328125 658.954833984375 251.89111328125 C 659.1791381835938 251.89111328125 659.361083984375 252.0730590820312 659.361083984375 252.2974243164062 L 659.361083984375 259.40673828125 C 659.361083984375 259.6310729980469 659.1791381835938 259.8130798339844 658.954833984375 259.8130798339844 C 658.7305297851562 259.8130798339844 658.5485229492188 259.6310729980469 658.5485229492188 259.40673828125 L 658.5485229492188 252.2974243164062 Z" fill="#ffffff" stroke="none" stroke-width="1" stroke-miterlimit="10" stroke-linecap="butt" /></svg>';
-const String _svg_gow1z9 =
-    '<svg viewBox="144.4 153.4 12.2 13.0" ><path transform="translate(-128.01, -69.16)" d="M 283.7218017578125 228.6105804443359 L 282.1484680175781 228.6105804443359 L 282.1484680175781 227.059326171875 C 282.1484680175781 226.4163208007812 281.6272888183594 225.8951568603516 280.9843139648438 225.8951568603516 L 278.9290771484375 225.8951568603516 L 278.9290771484375 225.0331268310547 L 282.3063659667969 223.7477874755859 C 282.3934936523438 223.6926574707031 282.3846130371094 223.5628967285156 282.2908325195312 223.5200805664062 L 278.7328796386719 222.5514526367188 C 278.7210693359375 222.5460510253906 278.708984375 222.5430450439453 278.697021484375 222.5413513183594 C 278.6739501953125 222.5357666015625 278.6502380371094 222.5319671630859 278.6253967285156 222.5319671630859 L 278.5140380859375 222.5319671630859 C 278.3463134765625 222.5319671630859 278.2103576660156 222.6679077148438 278.2103576660156 222.8356018066406 L 278.2103576660156 225.8951568603516 L 276.1551513671875 225.8951568603516 C 275.5121459960938 225.8951568603516 274.9908752441406 226.4163208007812 274.9908752441406 227.059326171875 L 274.9908752441406 228.6105804443359 L 273.4175720214844 228.6105804443359 C 272.8815002441406 228.6105804443359 272.4469909667969 229.0451812744141 272.4469909667969 229.5811920166016 L 272.4469909667969 234.5613708496094 C 272.4469909667969 235.0974578857422 272.8815002441406 235.5319671630859 273.4175720214844 235.5319671630859 L 276.1551513671875 235.5319671630859 L 277.2015991210938 235.5319671630859 L 277.2015991210938 232.5186614990234 C 277.2015991210938 231.9920196533203 277.6285095214844 231.565185546875 278.1551818847656 231.565185546875 L 278.9842834472656 231.565185546875 C 279.5108947753906 231.565185546875 279.9378356933594 231.9920196533203 279.9378356933594 232.5186614990234 L 279.9378356933594 235.5319671630859 L 280.9843139648438 235.5319671630859 L 283.7218017578125 235.5319671630859 C 284.2578735351562 235.5319671630859 284.6924438476562 235.0974578857422 284.6924438476562 234.5613708496094 L 284.6924438476562 229.5811920166016 C 284.6924438476562 229.0451812744141 284.2578735351562 228.6105804443359 283.7218017578125 228.6105804443359 Z" fill="#ffffff" stroke="none" stroke-width="1" stroke-miterlimit="10" stroke-linecap="butt" /></svg>';
-const String _svg_4lxdll =
-    '<svg viewBox="143.7 151.0 14.0 16.7" ><path transform="matrix(-0.866025, -0.5, 0.5, -0.866025, 144.67, 167.02)" d="M 0.7526724934577942 0.5132254958152771 L 0.6626434326171875 0.1666281670331955 C 0.6169024705886841 -0.00942822452634573 0.3917905688285828 -0.05911920219659805 0.2768227458000183 0.081474669277668 L 0.05055668950080872 0.3582927286624908 C -0.05452590063214302 0.4868596196174622 0.01261784136295319 0.6815704107284546 0.1742850989103317 0.7172239422798157 L 0.4906101524829865 0.7870519161224365 C 0.6523373126983643 0.822701632976532 0.794459342956543 0.6741995811462402 0.7526724934577942 0.5132254958152771 Z" fill="#ffffff" stroke="none" stroke-width="1" stroke-miterlimit="10" stroke-linecap="butt" /><path transform="matrix(-0.866025, -0.5, 0.5, -0.866025, 146.17, 167.35)" d="M 1.180817008018494 0.08138419687747955 L 0.05055016279220581 1.464014053344727 C -0.05453242734074593 1.592577219009399 0.01262629311531782 1.787280678749084 0.1742935478687286 1.822934031486511 L 1.405340194702148 2.09459662437439 L 2.636357069015503 2.366206169128418 C 2.79802417755127 2.401855945587158 2.940221071243286 2.253354072570801 2.898389339447021 2.092379808425903 L 2.448708295822144 0.361145406961441 C 2.427680253982544 0.2801172733306885 2.363172292709351 0.2177999168634415 2.28179931640625 0.1998233944177628 L 1.399683713912964 0.005174510180950165 C 1.318251013755798 -0.01275707501918077 1.233702898025513 0.01666264235973358 1.180817008018494 0.08138419687747955 Z" fill="#ffffff" stroke="none" stroke-width="1" stroke-miterlimit="10" stroke-linecap="butt" /><path transform="matrix(-0.866025, -0.5, 0.5, -0.866025, 149.82, 167.74)" d="M 2.909980535507202 15.64735126495361 L 0.1741584837436676 15.04369068145752 C 0.05457803606987 15.01731491088867 -0.02108810096979141 14.89852142333984 0.005167783703655005 14.77836418151855 L 3.196837902069092 0.1750389635562897 C 3.223123788833618 0.05486711487174034 3.341341495513916 -0.02115136571228504 3.460952043533325 0.005216189660131931 L 4.828847885131836 0.3070839047431946 L 6.196759223937988 0.6088917255401611 C 6.316369533538818 0.6352667212486267 6.392020702362061 0.7540529370307922 6.365735054016113 0.8742173314094543 L 3.174049615859985 15.47752857208252 C 3.147808790206909 15.59770011901855 3.029576063156128 15.67372608184814 2.909980535507202 15.64735126495361 Z" fill="#ffffff" stroke="none" stroke-width="1" stroke-miterlimit="10" stroke-linecap="butt" /><path transform="matrix(-0.866025, -0.5, 0.5, -0.866025, 149.34, 167.34)" d="M 0.5146351456642151 15.14903354644775 L 0.01431138720363379 15.03863620758057 C -0.007585735060274601 15.03382015228271 -0.004020837135612965 14.93251991271973 0.02227998152375221 14.81235599517822 L 3.213935375213623 0.2089781314134598 C 3.240191221237183 0.08881375938653946 3.279192924499512 -0.004638311453163624 3.301090002059937 0.0001694052771199495 L 3.551229476928711 0.05537577718496323 L 3.801413774490356 0.1105671674013138 C 3.82326602935791 0.1153823733329773 3.819746017456055 0.2166825532913208 3.793445348739624 0.3368543982505798 L 0.6017748713493347 14.94022464752197 C 0.575533926486969 15.06034469604492 0.5365172624588013 15.15384197235107 0.5146351456642151 15.14903354644775 Z" fill="#ffdbba" stroke="none" stroke-width="1" stroke-miterlimit="10" stroke-linecap="butt" /><path transform="matrix(-0.866025, -0.5, 0.5, -0.866025, 148.34, 166.42)" d="M 0.5146054625511169 15.14901828765869 L 0.01428172457963228 15.0386209487915 C -0.007555487100034952 15.03380584716797 -0.004035521764308214 14.93249034881592 0.02225031889975071 14.81233310699463 L 3.213920593261719 0.2089780420064926 C 3.240176439285278 0.08885860443115234 3.279178142547607 -0.004638399463146925 3.301075220108032 0.0001693175581749529 L 3.551214694976807 0.05536819994449615 L 3.801398992538452 0.1105670779943466 C 3.823236227035522 0.1153747960925102 3.819716453552246 0.2167348861694336 3.79343056678772 0.3368468284606934 L 0.60176020860672 14.94020175933838 C 0.5755192637443542 15.06037425994873 0.5365025997161865 15.15382671356201 0.5146054625511169 15.14901828765869 Z" fill="#ffdbba" stroke="none" stroke-width="1" stroke-miterlimit="10" stroke-linecap="butt" /></svg>';
