+++ flutter/safeway_direction/lib/main.dart
@@ -1,14 +1,12 @@
 import 'package:flutter/material.dart';
-import 'iPhoneXXSpage221/iPhoneXXSmainpage5.dart';
-import 'iPhoneXXSpage221/iPhoneXXSpage21.dart';
-
+import 'firstPage.dart';
 void main() => runApp(KidsKeeper());
 
 class KidsKeeper extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      home : iPhoneXXSmainpage5(),
+      home : firstPage(),
     );
   }
 }
